requirements.txt
asgiref==3.4.1
Django==4.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
mysqlclient==2.1.0
PyJWT==1.7.1
pytz==2021.3
sqlparse==0.4.2
